Declara o início, o intervalo de duração e o fim dos episódios de mudança de regime

Declara o início, o intervalo de duração e o fim dos episódios de mudança de regime

Usage

epi_progress(.data, .define_epi, .measure)

Arguments

.data

dados do projeto "Varieties of Democracy".

.define_epi

Lista de parâmetros que delimitam a extensão dos episódios de mudança de regime.

.stop: intervalo temporal que encerra os episódios de mudança de regime (em anos).

Por definição, quando a variação no índice de democracia for < 0.01 por três anos consecutivos em casos de democratização, e > -0.01 em casos de autocratização, .stop encerra os episódios de mudança de regime. Durante os episódios, variações menores que o .threshold indicam: 1) eventos de reação às alterações no status quo ou 2) a escasez de recursos para a implementação de novas alterações no regime.

.keep: número de eventos de estagnação ou reação aceitos durante episódios de mudança. Por definição, dois eventos.

.cumulative: define o valor acumulado mínimo para codificar um episódio de mudança de regime.

Por definição, 10% (0,1) em democratizações e -10% (-0,1) em autocratizações.

.after: quantidade máxima de anos em que o índice de democracia poderá não atingir o ponto de corte e ainda assim ter esses anos incluídos no calculo dos episódios de mudança de regime. Depende de .after_value.

.before: quantidade máxima de anos em que o índice de democracia poderá não atingir o ponto de corte, mas ainda assim ter esses anos incluídos no calculo dos episódios de mudança de regime. Depende de .before_value.

.after_value: variação mínima, inferior ao .threshold, para extensão do período de encerramento dos episódios.

.before_value: variação mínima, inferior ao .threshold, para extensão do período de abertura dos episódios.

.measure

"autocratization" ou "democratization"