Cria a tipologia Episódios de Autocratização e Democratização

typology detecta e classifica seis tipos de mudança de regime. O algorítmo considera explicitamente que episódios de mudança de regime são interpostos por eventos de reação.

Para episódios de autocratização, codifica-se casos de:

Aprofundamento autoritário: indica que instituições políticas em autocracias perderam atributos democráticos

Colapso democrático: designa casos de substituição de regimes democráticos por regimes autoritários

Retrocesso democrático: indica episódios de colapso democrático malsucedidos, isto é, casos em que a coalizão de atores antidemocráticos não obtiveram sucesso em romper o regime.

Para episódios de democratização, codifica-se casos de:

Aprofundamento democrático: indica que novos atributos democráticos foram incorporados em instituições políticas.

Transição democrática: designa casos de substituição de regimes autocráticos por regimes democráticos

Liberalização: indica episódios de transições democráticas malsucedidas, isto é, casos em que a coalizão de atores democráticos não obtiveram sucesso em encerrar regimes autoritários.

A tipologia baseia-se na generalização do terceiro pressuposto da proposta teórica que desenvolvi no capítulo três de minha dissertação para o exame de casos de retrocesso democrático.

Usage

epi_build(
  .data,
  ...,
  .index,
  .threshold,
  .define_epi,
  .events,
  .measure,
  .row_update = TRUE
)

typology(
  .data,
  ...,
  .index = IDE,
  .aut_parameters = list(.threshold = -0.01, .define_epi = list(.stop = 3, .keep = 2,
    .cumulative = -0.1, .before = 1, .after = 1, .before_value = -0.009, .after_value =
    -0.009), .events = list(.forward = c(-0.0099, -0.006), .stagnant = c(-0.0059,
    0.0059), .backward = 0.0059)),
  .dem_parameters = list(.threshold = 0.01, .define_epi = list(.stop = 3, .keep = 2,
    .cumulative = 0.1, .before = 1, .after = 1, .before_value = 0.009, .after_value =
    0.009), .events = list(.forward = c(0.006, 0.0099), .stagnant = c(-0.0059, 0.0059),
    .backward = -0.0059)),
  .row_update = TRUE,
  .all_vars = FALSE,
  .add_coups = TRUE
)

Arguments

.data

dados do projeto "Varieties of Democracy".

...

Variáveis que devem ser especificadas

.year = série histórica

.countries = países

.typology_d = classificação dicotômica

.typology_p = classificação policotômica

.reg_info = nome do regime, quando ele começa e quando ele termina

.reg_end_type = nome do processo mais importante que resultou no colapso do regime

.index

Índice de democracia usado para estimar os episódios de mudança de regime.

IDE: Índice de Democracia Eleitoral

IDL: Índice de Democracia Liberal

IDI: Índice de Democracia Igualitária

IDP: Índice de Democracia Participativa

IDD: Índice de Democracia Deliberativa

Por definição, eu adoto o Índice de Democracia Eleitoral (IDE), que mensura os pré-resiquitos institucionais da democracia, conforme foi proposto por Dahl (1971) e revisto em Dahl(1989), Dahl (1998) e em Coppedge et al.(2020).

IDE corresponde ao princípio, valor ou aspecto eleitoral da democracia. Os demais índices agregam o Índice de Democracia Eleitoral a componentes específicos da democracia. Por exemplo, o Índice de Democracia Liberal equilave ao: IDE + componente liberal da democracia.

Para mais informações, consulte o guia do usuário do projeto v-DEM (Coppedge et al., 2020) e o codebook do banco de dados (Coppedge et al., 2022).

.threshold

Define o valor mínimo de variação no índice de democracia para determinar o ínicio de processos de mudança de regime. .threshold aceita valores entre 0 e 1.

Por definição, 0.01(1%) em episódios de democratização e -0.01(-1%) em episódios de autocratização.

.define_epi

Lista de parâmetros que delimitam a extensão dos episódios de mudança de regime.

.stop: intervalo temporal que encerra os episódios de mudança de regime (em anos).

Por definição, quando a variação no índice de democracia for < 0.01 por três anos consecutivos em casos de democratização, e > -0.01 em casos de autocratização, .stop encerra os episódios de mudança de regime. Durante os episódios, variações menores que o .threshold indicam: 1) eventos de reação às alterações no status quo ou 2) a escasez de recursos para a implementação de novas alterações no regime.

.keep: número de eventos de estagnação ou reação aceitos durante episódios de mudança. Por definição, dois eventos.

.cumulative: define o valor acumulado mínimo para codificar um episódio de mudança de regime.

Por definição, 10% (0,1) em democratizações e -10% (-0,1) em autocratizações.

.after: quantidade máxima de anos em que o índice de democracia poderá não atingir o ponto de corte e ainda assim ter esses anos incluídos no calculo dos episódios de mudança de regime. Depende de .after_value.

.before: quantidade máxima de anos em que o índice de democracia poderá não atingir o ponto de corte, mas ainda assim ter esses anos incluídos no calculo dos episódios de mudança de regime. Depende de .before_value.

.after_value: variação mínima, inferior ao .threshold, para extensão do período de encerramento dos episódios.

.before_value: variação mínima, inferior ao .threshold, para extensão do período de abertura dos episódios.

.events

define os eventos de avanço, estagnação ou reação.

.forward:

por definição -0.0099 <= .forward <= -0.006 em episódios de autocratização e 0.006 <= .forward <= 0.0099 em episódios de democratização. .forward especifica variações menores, mas próximas ao threshold.

.stagnant:

por definição -0.0059 <= .stagnant <= 0.0059 em episódios de autocratização e de democratização. .stagnant especifica eventos de estagnação.

.backward:

por definição, .backward < 0.0059 em episódios de autocratização e .backward < - 0.0059 em episódios de democratização. .backward, por fim, indica eventos de reação.

.measure

"autocratization" ou "democratization"

.row_update

desconsidera as transições que resultam de variações anuais em IDE contidas no intervalo [-0.05; 0.05] para todos os casos em que tais episódios sucedem imediatamente um retorno ao regime anterior. Em outros termos, transições de regime com duração de um ano, fruto de variações de |5%| em IDE são descartadas. Argumento que essas transições devem ser tratadas como ruídos, pois o próprio conceito de regime político pressupõe um mínimo de continuidade no tempo (Schedler, 2013). Por definição, TRUE.

Essa operação reduz o número de episódios identificados pela tipologia row.

.aut_parameters,

Veja os parâmetros .threshold, .define_epi e .events

.dem_parameters

Veja os parâmetros .threshold, .define_epi e .events

.all_vars

Se TRUE, retorna dois bancos de dados adicionais com todas as variáveis usadas para construir os episódios de democratização e autocratização.

.add_coups

Adiciona episódios de golpes e autogolpes à tipologia

Note

A tipologia usa dados do projeto Varieties of Democracy (v-DEM) e é inspirada nos estudos de Lührmann et al. (2018), Lührmann e Lindberg (2019), Mainwaring e Bizzaro (2019), Wilson et al. (2020), Maerz et al. (2021) e Haggard e Kaufman (2021); especialmente nas tipologias Regimes of the World (row) e Episodes of Regime Transformation (ERT).

Examples

library(ead)

typology(
  vdem,
  .countries = country_name,
  .year = year,
  .typology_p = regime,
  .typology_d = regime2,
  .reg_info = reg_info,
  .reg_end_type = reg_end_type,
  .index = IDE,
  .add_coups = TRUE,
  .all_vars = TRUE
) -> typology
#> ℹ This R version contain the native pipe.
#> ℹ R version: 4.2.2
#> ⠙ Identificando os episódios...
#> ───────────────────────────────  autocratization  ──────────────────────────────
#> ⠙ Criando as medidas...
#> ⠹ Criando as medidas...
#> ✔ Criando as medidas... [1.7s]
#> 
#> ⠙ Identificando os episódios...
#> ⠹ Identificando os episódios...
#> ✔ Identificando os episódios... [5.4s]
#> 
#> ⠙ Atualizando 'Regimes of the World'...
#> ⠹ Atualizando 'Regimes of the World'...
#> ✔ Atualizando 'Regimes of the World'... [891ms]
#> 
#> ⠙ Iniciando a codificação...
#> ⠹ Iniciando a codificação...
#> ✔ Codificação concluída com sucesso! [4.9s]
#> 
#> ⠙ Identificando os episódios...
#> ───────────────────────────────  democratization  ──────────────────────────────
#> ⠙ Criando as medidas...
#> ✔ Criando as medidas... [1.4s]
#> 
#> ⠙ Identificando os episódios...
#> ⠹ Identificando os episódios...
#> ✔ Identificando os episódios... [6.5s]
#> 
#> ⠙ Atualizando 'Regimes of the World'...
#> ✔ Atualizando 'Regimes of the World'... [778ms]
#> 
#> ⠙ Iniciando a codificação...
#> ⠹ Iniciando a codificação...
#> ✔ Codificação concluída com sucesso! [6.7s]
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> 
#>                          ┌──────────────────────────────┐
#>                          │                              │
#>                          │   A Tipologia está pronta!   │
#>                          │                              │
#>                          └──────────────────────────────┘
#> 
#> ℹ Você construiu a tipologia com as seguintes especificações:
#> Em episódios de autocratização:
#> 
#> • threshold = -0.01
#> • keep = 2
#> • stop = 3
#> • cumulative = -0.1
#> Em episódios de democratização:
#> 
#>   • threshold = 0.01
#>   • keep = 2
#>   • stop = 3
#>   • cumulative = 0.1
#> Índice de democracia:
#> 
#>   • IDE